Introduce system.Dialer interface, which is capable of establishing a capnp RPC connection.
Breaks system package.  TODO:  implement system.Bootstrap.
